requirements.txt
python-dotenv==0.19.2
requests==2.26.0
terminaltables==3.1.7

